feat: Separate service and implementation
