+++ src/main/java/com/psj/itembrowser/order/service/impl/ShippingFeeService.java
@@ -1,19 +1,16 @@
 package com.psj.itembrowser.order.service.impl;
 
-import org.springframework.stereotype.Service;
-
 /**
  *packageName    : com.psj.itembrowser.order.service.impl
  * fileName       : ShippingFeeService
  * author         : ipeac
- * date           : 2024-02-12
+ * date           : 2024-02-13
  * description    :
  * ===========================================================
  * DATE              AUTHOR             NOTE
  * -----------------------------------------------------------
- * 2024-02-12        ipeac       최초 생성
+ * 2024-02-13        ipeac       최초 생성
  */
-@Service
-public class ShippingFeeService {
-
+public interface ShippingFeeService {
+	double calculateShippingFee(double totalPrice);
 }
